Moved test_strategy.ts to test directory
